terraform: Move the terraform wrapper code in pkg/terraform
