+++ pkg/terraform/BUILD.bazel
@@ -1,0 +1,12 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = [
+        "doc.go",
+        "executor.go",
+        "terraform.go",
+    ],
+    importpath = "github.com/openshift/installer/pkg/terraform",
+    visibility = ["//visibility:public"],
+)
