+++ pkg/terraform/doc.go
@@ -1,0 +1,4 @@
+// Package terraform contains the utilities that's used for invoking
+// terraform executable under the given directory with the given
+// templates.
+package terraform
